Refactored some stuff with the blog posts.
