+++ package.json
@@ -8,8 +8,10 @@
     "axios": "^0.18.0",
     "babel-plugin-transform-regenerator": "^6.26.0",
     "babel-plugin-transform-runtime": "^6.23.0",
+    "browserslist": "^4.4.2",
     "bufferutil": "^3.0.5",
     "bulma": "^0.6.0",
+    "caniuse-lite": "^1.0.30000945",
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
     "gatsby": "^1.9.277",
