+++ src/components/blog/styles/BlogPostTemplate.sass
@@ -23,6 +23,11 @@
       line-height: 1.8em;
       @media (max-width: 500px)
         font-size: 18px;
+    blockquote
+      p 
+        font-style: italic;
+        opacity: 0.5;
+        font-family: 'Roboto';
   .post-description
       font-size: 21px;
       line-height: 1.8em;
