--- src/styles/Blog/BlogPostPage.module.css
@@ -1,66 +1,0 @@
-.container {
-    max-width: 827px;
-    margin: 0 auto;
-    padding-left: 1em;
-    padding-right: 1em;
-    padding-bottom: 1em;
-  }
-  
-  .header {
-    width: 100%;
-    margin-top: 4em;
-  }
-  
-  .header>h1 {
-    margin: 0;
-    padding-bottom: 6px;
-  }
-  
-  .subHeader {
-      font-size: 19px;
-    color: #a8a8a8;
-    margin-bottom: 1em;
-    text-transform: capitalize;
-  }
-  
-  .headerImageContainer {
-    text-align: center;
-  }
-  
-  .headerImage {
-    width: 100%;
-    height: 500px;
-    background-position-y: 50%;
-    background-position-x: 50%;
-  }
-  
-  .content p,
-  a,
-  span,
-  ol,
-  li,
-  ul {
-    font-size: 18px;
-    line-height: 1.7em;
-  }
-  
-  .content blockquote > p {
-        font-weight: bold;
-      font-style: italic;
-    font-size: 30px;
-      color: #a2a2a2;
-      border-left: solid 3px;
-      padding-left: 19px;
-  }
-  
-  @media (max-width: 600px) {
-    .header h1 {
-      font-size: 7vw;
-    }
-    .content h1,
-    h2,
-    h3,
-    h4 {
-      font-size: 30px;
-    }
-  }
